implemented first few steps to get the impl class to build a valid URL given a list of dependencies and then creating a Document object off of that
